ehs_backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ehs-backend"
version = "1.0.0"
description = "EHS Inspection Platform Backend"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pgvector>=0.3.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # ML — DistilBERT
    "torch>=2.4.0",
    "transformers>=4.45.0",
    "sentence-transformers>=3.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",

    # LangChain / RAG
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",

    # Azure OpenAI
    "openai>=1.54.0",

    # Groq (risk assessment chatbot)
    "groq>=0.11.0",

    # Document parsing (RA generation from PDF/DOCX)
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",

    # NLP (semantic chunking)
    "spacy>=3.7.0",

    # Validation / settings
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Logging
    "structlog>=24.4.0",

    # Utils
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "factory-boy>=3.3.0",
]
all = ["ehs-backend[dev]"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "scripts*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
